+++ api/scheduler/src/main/java/com/example/scheduler/DTOs/UserDTO.java
@@ -4,23 +4,23 @@ import java.util.Objects;
 
 public class UserDTO {
 
-    private int id;
+    private Long id;
     private String username;
     private String name;
     private String email;
 
     public UserDTO()  {}
 
-    public UserDTO(int id, String username, String name, String email){
+    public UserDTO(Long id, String username, String name, String email){
         this.id = id;
         this.username = username;
         this.name = name;
         this.email = email;
     }
 
-    public int getId() { return id;}
+    public Long getId() { return id;}
 
-    public void setId(int id) { this.id = id;}
+    public void setId(Long id) { this.id = id;}
 
     public String getUsername() { return username;}
